requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.0.4
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.10.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.1
applaunchservices==0.3.0
appnope==0.1.4
argcomplete==3.3.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.2.2
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==24.4.2
bleach==6.1.0
blinker==1.7.0
boto3==1.34.114
botocore==1.34.114
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cookiecutter==2.6.0
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.8
dnspython==2.6.1
docstring-to-markdown==0.15
docutils==0.21.2
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
flake8==7.0.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.10.0
google-api-python-client==2.108.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-re2==1.1
googleapis-common-protos==1.63.0
grpcio==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imagesize==1.4.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.25.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jellyfish==1.0.4
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.2.1
lazy-object-proxy==1.10.0
limits==3.11.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==4.9.3
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.0
numpydoc==1.7.0
oauthlib==3.2.2
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.47
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.2.5
pylint-venv==3.0.3
pyls-spyder==0.4.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreServices==10.3.1
pyobjc-framework-FSEvents==10.3.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.14
pytest==8.2.2
pytest-mock==3.14.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
QDarkStyle==3.2.3
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.5.2
QtPy==2.4.1
redis==5.0.4
referencing==0.35.0
requests==2.31.0
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rope==1.13.0
rpds-py==0.18.0
rsa==4.9
Rtree==1.3.0
s3transfer==0.10.1
setproctitle==1.3.3
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.5
spyder-kernels==2.5.2
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.2
three-merge==0.1.1
time-machine==2.14.1
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.13.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
whatthepatch==1.0.5
wrapt==1.16.0
WTForms==3.1.2
wurlitzer==3.1.1
XlsxWriter==3.1.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
